feat: refactor cache functions
