+++ apps/api/src/app/auth/usecases/password-reset/password-reset.usecase.ts
@@ -6,7 +6,7 @@ import { PasswordResetCommand } from './password-reset.command';
 import { ApiException } from '../../../shared/exceptions/api.exception';
 import { AuthService } from '../../services/auth.service';
 import { InvalidateCacheService } from '../../../shared/services/cache';
-import { entityBuilder } from '../../../shared/services/cache/keys';
+import { buildUserKey } from '../../../shared/services/cache/key-builders/entities';
 
 @Injectable()
 export class PasswordReset {
@@ -29,7 +29,7 @@ export class PasswordReset {
     const passwordHash = await bcrypt.hash(command.password, 10);
 
     await this.invalidateCache.invalidateByKey({
-      key: entityBuilder().user({
+      key: buildUserKey({
         _id: user._id,
       }),
     });
