+++ apps/api/src/app/events/usecases/send-message/send-message.usecase.ts
@@ -37,13 +37,9 @@ import {
 import { ANALYTICS_SERVICE } from '../../../shared/shared.module';
 import { ApiException } from '../../../shared/exceptions/api.exception';
 import { CachedEntity } from '../../../shared/interceptors/cached-entity.interceptor';
-import {
-  buildCommonKey,
-  CacheKeyPrefixEnum,
-  CacheKeyTypeEnum,
-  notificationTemplateQueryKeyBuild,
-} from '../../../shared/services/cache/keys';
 import { CachedQuery } from '../../../shared/interceptors/cached-query.interceptor';
+import { buildNotificationTemplateKey } from '../../../shared/services/cache/key-builders/queries';
+import { buildSubscriberKey } from '../../../shared/services/cache/key-builders/entities';
 
 @Injectable()
 export class SendMessage {
@@ -179,12 +175,9 @@ export class SendMessage {
 
   @CachedEntity({
     builder: (command: { subscriberId: string; _environmentId: string }) =>
-      buildCommonKey({
-        type: CacheKeyTypeEnum.ENTITY,
-        keyEntity: CacheKeyPrefixEnum.SUBSCRIBER,
-        environmentId: command._environmentId,
-        identifier: command.subscriberId,
-        identifierPrefix: 's',
+      buildSubscriberKey({
+        _environmentId: command._environmentId,
+        subscriberId: command.subscriberId,
       }),
   })
   private async getSubscriberBySubscriberId({
@@ -241,7 +234,7 @@ export class SendMessage {
 
   @CachedQuery({
     builder: (command: { _id: string; environmentId: string }) =>
-      notificationTemplateQueryKeyBuild().cache({
+      buildNotificationTemplateKey().cache({
         _environmentId: command.environmentId,
         identifiers: { id: command._id },
       }),
