+++ apps/api/src/app/integrations/usecases/create-integration/create-integration.usecase.ts
@@ -9,8 +9,8 @@ import { DeactivateSimilarChannelIntegrations } from '../deactivate-integration/
 import { CheckIntegrationCommand } from '../check-integration/check-integration.command';
 import { CheckIntegration } from '../check-integration/check-integration.usecase';
 import { ANALYTICS_SERVICE } from '../../../shared/shared.module';
-import { CacheKeyPrefixEnum, InvalidateCacheService } from '../../../shared/services/cache';
-import { buildCommonKey, CacheKeyTypeEnum } from '../../../shared/services/cache/keys';
+import { InvalidateCacheService } from '../../../shared/services/cache';
+import { buildIntegrationKey } from '../../../shared/services/cache/key-builders/entities';
 
 @Injectable()
 export class CreateIntegration {
@@ -46,11 +46,9 @@ export class CreateIntegration {
       }
 
       await this.invalidateCache.invalidateQuery({
-        key: buildCommonKey({
-          type: CacheKeyTypeEnum.QUERY,
-          keyEntity: CacheKeyPrefixEnum.INTEGRATION,
-          environmentId: command.environmentId,
-          identifier: command.userId,
+        key: buildIntegrationKey({
+          _environmentId: command.environmentId,
+          _id: command.userId,
         }),
       });
