+++ apps/api/src/app/shared/services/cache/index.ts
@@ -1,3 +1,2 @@
 export * from './invalidate-cache.service';
 export * from './cache.service';
-export { CacheKeyPrefixEnum } from './keys';
