+++ apps/api/src/app/subscribers/usecases/remove-subscriber/remove-subscriber.usecase.ts
@@ -4,7 +4,7 @@ import { RemoveSubscriberCommand } from './remove-subscriber.command';
 import { GetSubscriber } from '../get-subscriber';
 import { ApiException } from '../../../shared/exceptions/api.exception';
 import { InvalidateCacheService } from '../../../shared/services/cache';
-import { entityBuilder } from '../../../shared/services/cache/keys';
+import { buildSubscriberKey } from '../../../shared/services/cache/key-builders/entities';
 
 @Injectable()
 export class RemoveSubscriber {
@@ -24,7 +24,7 @@ export class RemoveSubscriber {
       });
 
       await this.invalidateCache.invalidateByKey({
-        key: entityBuilder().subscriber({
+        key: buildSubscriberKey({
           subscriberId: command.subscriberId,
           _environmentId: command.environmentId,
         }),
